+++ packages/merge-on-green/test/merge-on-green.ts
@@ -17,9 +17,8 @@ import {Probot, createProbot, ProbotOctokit} from 'probot';
 import {resolve} from 'path';
 import nock from 'nock';
 import sinon, {SinonStub} from 'sinon';
-import {describe, it, beforeEach, afterEach} from 'mocha';
+import {describe, it, beforeEach, afterEach, suite} from 'mocha';
 import handler from '../src/merge-on-green';
-import {CheckStatus, Reviews, Comment} from '../src/merge-logic';
 import {logger} from 'gcf-utils';
 import assert from 'assert';
 // eslint-disable-next-line node/no-extraneous-import
@@ -28,15 +27,6 @@ import {config} from '@probot/octokit-plugin-config';
 const TestingOctokit = ProbotOctokit.plugin(config);
 const testingOctokitInstance = new TestingOctokit({auth: 'abc123'});
 const sandbox = sinon.createSandbox();
-
-interface HeadSha {
-  sha: string;
-}
-
-interface CheckRuns {
-  name: string;
-  conclusion: string;
-}
 
 interface PR {
   number: number;
@@ -54,798 +44,92 @@ nock.disableNetConnect();
 
 const fixturesPath = resolve(__dirname, '../../test/Fixtures');
 
-function getReviewsCompleted(response: Reviews[]) {
-  return nock('https://api.github.com')
-    .get('/repos/testOwner/testRepo/pulls/1/reviews')
-    .reply(200, response);
-}
-
-function getLatestCommit(response: HeadSha[]) {
-  return nock('https://api.github.com')
-    .get('/repos/testOwner/testRepo/pulls/1/commits?per_page=100&page=1')
-    .reply(200, response);
-}
-
-function getStatusi(ref: string, response: CheckStatus[]) {
-  return nock('https://api.github.com')
-    .get(`/repos/testOwner/testRepo/commits/${ref}/statuses`)
-    .reply(200, response);
-}
-
-function getRuns(ref: string, response: CheckRuns) {
-  return nock('https://api.github.com')
-    .get(`/repos/testOwner/testRepo/commits/${ref}/check-runs`)
-    .reply(200, response);
-}
-
-function getCommentsOnPr(response: Comment[]) {
-  return nock('https://api.github.com')
-    .get('/repos/testOwner/testRepo/issues/1/comments')
-    .reply(200, response);
-}
-
-function removeMogLabel(label: string) {
-  return nock('https://api.github.com')
-    .delete(`/repos/testOwner/testRepo/issues/1/labels/${label}`)
-    .reply(200);
-}
-
-function merge() {
-  return nock('https://api.github.com')
-    .put('/repos/testOwner/testRepo/pulls/1/merge')
-    .reply(200, {sha: '123', merged: true, message: 'in a bottle'});
-}
-
-function dismissReview(reviewNumber: number) {
-  return nock('https://api.github.com')
-    .put(`/repos/testOwner/testRepo/pulls/1/reviews/${reviewNumber}/dismissals`)
-    .reply(200);
-}
-
-function mergeWithError() {
-  return nock('https://api.github.com')
-    .put('/repos/testOwner/testRepo/pulls/1/merge')
-    .reply(400);
-}
-
-function commentOnPR() {
-  return nock('https://api.github.com')
-    .post('/repos/testOwner/testRepo/issues/1/comments')
-    .reply(200);
-}
-
-function updateBranch() {
-  return nock('https://api.github.com')
-    .put('/repos/testOwner/testRepo/pulls/1/update-branch')
-    .reply(200);
-}
-
 function getBranchProtection(status: number, requiredStatusChecks: string[]) {
-  return nock('https://api.github.com')
-    .get('/repos/testOwner/testRepo/branches/master/protection')
-    .reply(status, {
-      required_status_checks: {
-        contexts: requiredStatusChecks,
-      },
-    });
-}
-
-function getPRCleanUp(state: string, merged: boolean) {
-  return nock('https://api.github.com')
-    .get('/repos/testOwner/testRepo/pulls/1')
-    .reply(200, {state, merged});
-}
-
-function getLabels(name: string) {
-  return nock('https://api.github.com')
-    .get('/repos/testOwner/testRepo/issues/1/labels')
-    .reply(200, [{name}]);
-}
-
-function getRateLimit(remaining: number) {
-  return nock('https://api.github.com')
-    .get('/rate_limit')
-    .reply(200, {
-      resources: {
-        core: {
-          limit: 5000,
-          remaining: remaining,
-          reset: 1372700873,
+    return nock('https://api.github.com')
+      .get('/repos/testOwner/testRepo/branches/master/protection')
+      .reply(status, {
+        required_status_checks: {
+          contexts: requiredStatusChecks,
         },
-      },
-    });
-}
-
-function removeReaction() {
-  return nock('https://api.github.com')
-    .delete('/repos/testOwner/testRepo/issues/1/reactions/1')
-    .reply(204);
-}
-
-function getPR(
-  mergeable: boolean,
-  mergeableState: string,
-  state: string,
-  labels: {name: string}[] = []
-) {
-  return nock('https://api.github.com')
-    .get('/repos/testOwner/testRepo/pulls/1')
-    .reply(200, {
-      title: 'Test PR',
-      body: 'Test Body',
-      state,
-      mergeable,
-      mergeable_state: mergeableState,
-      user: {login: 'login'},
-      labels,
-    });
-}
-
+      });
+  }
+  
+  function getPRCleanUp(state: string, merged: boolean) {
+    return nock('https://api.github.com')
+      .get('/repos/testOwner/testRepo/pulls/1')
+      .reply(200, {state, merged});
+  }
+  
+  function getLabels(name: string) {
+    return nock('https://api.github.com')
+      .get('/repos/testOwner/testRepo/issues/1/labels')
+      .reply(200, [{name}]);
+  }
+  
 function searchForPRs(pr: PR[], labelName: string) {
-  return nock('https://api.github.com')
-    .get(
-      `/search/issues?q=is%3Aopen%20is%3Apr%20user%3Agoogleapis%20label%3A%22${labelName}%22`
-    )
-    .reply(200, pr);
-}
+    return nock('https://api.github.com')
+      .get(
+        `/search/issues?q=is%3Aopen%20is%3Apr%20user%3Agoogleapis%20label%3A%22${labelName}%22`
+      )
+      .reply(200, pr);
+  }
 
-//meta-note about the schedule.repository as any; currently GH does not support this type, see
-//open issue for a fix: https://github.com/octokit/webhooks.js/issues/277
-describe('merge-on-green', () => {
-  let probot: Probot;
-  const loggerStub = sandbox.stub(logger, 'error').throwsArg(0);
+  function removeMogLabel(label: string) {
+    return nock('https://api.github.com')
+      .delete(`/repos/testOwner/testRepo/issues/1/labels/${label}`)
+      .reply(200);
+  }
 
-  beforeEach(() => {
-    probot = createProbot({
-      githubToken: 'abc123',
-      Octokit: ProbotOctokit.defaults({
-        retry: {enabled: false},
-        throttle: {enabled: false},
-      }),
+
+  function commentOnPR() {
+    return nock('https://api.github.com')
+      .post('/repos/testOwner/testRepo/issues/1/comments')
+      .reply(200);
+  }
+  
+
+
+  function removeReaction() {
+    return nock('https://api.github.com')
+      .delete('/repos/testOwner/testRepo/issues/1/reactions/1')
+      .reply(204);
+  }
+// general structure of tests: this file tests the merge-on-green logic, 
+// which wraps the merge logic itself. I have attempted to divide up the
+// tests based on what its testing, but also based on sandbox scopes for
+// mocking the different GCP functions. You'll see suite blocks for the
+// different stubs for GCP functions, and describe blocks to describe the
+// functionality of what it's actually testing
+describe('merge-on-green wrapper logic', () => {
+    let probot: Probot;
+    let loggerStub: SinonStub;
+
+    before(() => {
+      loggerStub = sandbox.stub(logger, 'error').throwsArg(0);
+    })
+
+    after(() => {
+      loggerStub.restore();
+    })
+  
+    beforeEach(() => {
+      probot = createProbot({
+        githubToken: 'abc123',
+        Octokit: ProbotOctokit.defaults({
+          retry: {enabled: false},
+          throttle: {enabled: false},
+        }),
+      });
+  
+      const app = probot.load(handler);
+      app.auth = async () => testingOctokitInstance;
+    });
+  
+    afterEach(() => {
+      nock.cleanAll();
     });
 
-    const app = probot.load(handler);
-    app.auth = async () => testingOctokitInstance;
-  });
-
-  afterEach(() => {
-    nock.cleanAll();
-  });
-
-  describe('merge-logic', () => {
-    handler.removePR = async () => {
-      return Promise.resolve(undefined);
-    };
-
-    describe('with default/normal get Datastore payload', () => {
-      handler.getDatastore = async () => {
-        const pr = [
-          [
-            {
-              repo: 'testRepo',
-              number: 1,
-              owner: 'testOwner',
-              created: Date.now(),
-              branchProtection: ['Special Check'],
-              label: 'automerge',
-              author: 'testOwner',
-              reactionId: 1,
-              url: 'url/url',
-              installationId: 123456,
-            },
-          ],
-        ];
-        return pr;
-      };
-
-      it('merges a PR on green', async () => {
-        const scopes = [
-          getRateLimit(5000),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', [
-            {state: 'success', context: 'Special Check'},
-          ]),
-          getPR(true, 'clean', 'open'),
-          getCommentsOnPr([]),
-          merge(),
-          removeMogLabel('automerge'),
-          removeReaction(),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      it('fails when a review has not been approved', async () => {
-        const scopes = [
-          getRateLimit(5000),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-            {
-              user: {login: 'octokitten'},
-              state: 'CHANGES_REQUESTED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', [
-            {state: 'success', context: 'Special Check'},
-          ]),
-          getCommentsOnPr([]),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      it('fails if there is no commit', async () => {
-        const scopes = [
-          getRateLimit(5000),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getLatestCommit([]),
-          getStatusi('', [
-            {state: 'success', context: 'Kokoro - Test: Binary Compatibility'},
-          ]),
-          getCommentsOnPr([]),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      it('fails if there are no status checks', async () => {
-        const scopes = [
-          getRateLimit(5000),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', []),
-          getRuns('6dcb09b5b57875f334f61aebed695e2e4193db5e', {
-            name: '',
-            conclusion: '',
-          }),
-          getCommentsOnPr([]),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      it('fails if the status checks have failed', async () => {
-        const scopes = [
-          getRateLimit(5000),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', [
-            {state: 'failure', context: 'Special Check'},
-          ]),
-          getCommentsOnPr([]),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      it('passes if checks are actually check runs', async () => {
-        const scopes = [
-          getRateLimit(5000),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', []),
-          getRuns('6dcb09b5b57875f334f61aebed695e2e4193db5e', {
-            name: 'Special Check',
-            conclusion: 'success',
-          }),
-          getCommentsOnPr([]),
-          getPR(true, 'clean', 'open'),
-          merge(),
-          removeMogLabel('automerge'),
-          removeReaction(),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      it('fails if there is a do not merge label', async () => {
-        const scopes = [
-          getRateLimit(5000),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', []),
-          getRuns('6dcb09b5b57875f334f61aebed695e2e4193db5e', {
-            name: 'Special Check',
-            conclusion: 'success',
-          }),
-          getCommentsOnPr([]),
-          getPR(true, 'clean', 'open', [{name: 'do not merge'}]),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      it('fails if no one has reviewed the PR', async () => {
-        const scopes = [
-          getRateLimit(5000),
-          getReviewsCompleted([]),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', []),
-          getRuns('6dcb09b5b57875f334f61aebed695e2e4193db5e', {
-            name: 'Special Check',
-            conclusion: 'success',
-          }),
-          getCommentsOnPr([]),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      //This method is supposed to include an error
-      it('updates a branch if merge returns error and branch is behind', async () => {
-        loggerStub.restore();
-
-        const scopes = [
-          getRateLimit(5000),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', [
-            {state: 'success', context: 'Special Check'},
-          ]),
-          getCommentsOnPr([]),
-          getPR(true, 'behind', 'open'),
-          mergeWithError(),
-          updateBranch(),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      //This test is supposed to include an error
-      it('comments on PR if branch is dirty and merge returns with error', async () => {
-        loggerStub.restore();
-
-        const scopes = [
-          getRateLimit(5000),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', [
-            {state: 'success', context: 'Special Check'},
-          ]),
-          getCommentsOnPr([]),
-          getPR(true, 'dirty', 'open'),
-          mergeWithError(),
-          commentOnPR(),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      //This test is supposed to include an error
-      it('does not comment if comment is already on PR and merge errors', async () => {
-        loggerStub.restore();
-
-        const scopes = [
-          getRateLimit(5000),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', [
-            {state: 'success', context: 'Special Check'},
-          ]),
-          getCommentsOnPr([
-            {
-              body:
-                'Your PR has conflicts that you need to resolve before merge-on-green can automerge',
-            },
-          ]),
-          getPR(true, 'dirty', 'open'),
-          mergeWithError(),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      it('does not execute if there is no more space for requests', async () => {
-        // eslint-disable-next-line @typescript-eslint/no-var-requires
-        const scopes = [getRateLimit(0)];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-    });
-
-    describe('with different Datastore payloads', () => {
-      it('posts a comment on the PR if the flag is set to stop and the merge has failed', async () => {
-        handler.getDatastore = async () => {
-          const pr = [
-            [
-              {
-                repo: 'testRepo',
-                number: 1,
-                owner: 'testOwner',
-                created: -14254782000,
-                branchProtection: ['Special Check'],
-                label: 'automerge',
-                author: 'testOwner',
-                reactionId: 1,
-              },
-            ],
-          ];
-          return pr;
-        };
-
-        const scopes = [
-          getRateLimit(5000),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', [
-            {state: 'failure', context: 'Special Check'},
-          ]),
-          getCommentsOnPr([]),
-          commentOnPR(),
-          removeMogLabel('automerge'),
-          removeReaction(),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      it('rejects status checks that do not match the required check', async () => {
-        handler.getDatastore = async () => {
-          const pr = [
-            [
-              {
-                repo: 'testRepo',
-                number: 1,
-                owner: 'testOwner',
-                created: Date.now(),
-                branchProtection: ["this is what we're looking for"],
-                label: 'automerge',
-                author: 'testOwner',
-                reactionId: 1,
-              },
-            ],
-          ];
-          return pr;
-        };
-
-        const scopes = [
-          getRateLimit(5000),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          //Intentionally giving this status check a misleading name. We want subtests to match the beginning
-          //of required status checks, not the other way around. i.e., if the required status check is "passes"
-          //then it should reject a status check called "passe", but pass one called "passesS"
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', [
-            {state: 'success', context: "this is what we're looking fo"},
-          ]),
-          getRuns('6dcb09b5b57875f334f61aebed695e2e4193db5e', {
-            name: "this is what we're looking fo",
-            conclusion: 'success',
-          }),
-          getCommentsOnPr([]),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      it('accepts status checks that match the beginning of the required status check', async () => {
-        handler.getDatastore = async () => {
-          const pr = [
-            [
-              {
-                repo: 'testRepo',
-                number: 1,
-                owner: 'testOwner',
-                created: Date.now(),
-                branchProtection: ["this is what we're looking for"],
-                label: 'automerge',
-                author: 'testOwner',
-                reactionId: 1,
-              },
-            ],
-          ];
-          return pr;
-        };
-
-        const scopes = [
-          getRateLimit(5000),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', [
-            {
-              state: 'success',
-              context: "this is what we're looking for/subtest",
-            },
-          ]),
-          getCommentsOnPr([]),
-          getPR(true, 'clean', 'open'),
-          merge(),
-          removeMogLabel('automerge'),
-          removeReaction(),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      it('merges a PR on green with exact label', async () => {
-        handler.getDatastore = async () => {
-          const pr = [
-            [
-              {
-                repo: 'testRepo',
-                number: 1,
-                owner: 'testOwner',
-                created: Date.now(),
-                branchProtection: ['Special Check'],
-                label: 'automerge: exact',
-                author: 'testOwner',
-                reactionId: 1,
-              },
-            ],
-          ];
-          return pr;
-        };
-
-        const scopes = [
-          getRateLimit(5000),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '6dcb09b5b57875f334f61aebed695e2e4193db5e',
-              id: 12345,
-            },
-          ]),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', [
-            {state: 'success', context: 'Special Check'},
-          ]),
-          getPR(true, 'clean', 'open'),
-          getCommentsOnPr([]),
-          merge(),
-          removeMogLabel('automerge%3A%20exact'),
-          removeReaction(),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-
-      it('dismisses reviews if automerge label is set to exact', async () => {
-        handler.getDatastore = async () => {
-          const pr = [
-            [
-              {
-                repo: 'testRepo',
-                number: 1,
-                owner: 'testOwner',
-                created: Date.now(),
-                branchProtection: ['Special Check'],
-                label: 'automerge: exact',
-                author: 'testOwner',
-                reactionId: 1,
-              },
-            ],
-          ];
-          return pr;
-        };
-
-        const scopes = [
-          getRateLimit(5000),
-          getLatestCommit([{sha: '6dcb09b5b57875f334f61aebed695e2e4193db5e'}]),
-          getReviewsCompleted([
-            {
-              user: {login: 'octocat'},
-              state: 'APPROVED',
-              commit_id: '12345',
-              id: 12345,
-            },
-            {
-              user: {login: 'octokitten'},
-              state: 'APPROVED',
-              commit_id: '12346',
-              id: 12346,
-            },
-          ]),
-          dismissReview(12345),
-          dismissReview(12346),
-          getStatusi('6dcb09b5b57875f334f61aebed695e2e4193db5e', [
-            {state: 'success', context: 'Special Check'},
-          ]),
-          getCommentsOnPr([]),
-        ];
-
-        await probot.receive({
-          name: 'schedule.repository' as '*',
-          payload: {org: 'testOwner'},
-          id: 'abc123',
-        });
-
-        scopes.forEach(s => s.done());
-      });
-    });
-  });
-
-  describe('merge-on-green wrapper logic', () => {
-    describe('adding-a-PR-to-Datastore (addPR) method (no stub)', async () => {
+    describe('adding-a-PR-to-Datastore (addPR) method', async () => {
       let removePRStub: SinonStub;
       let getPRStub: SinonStub;
 
@@ -859,7 +143,6 @@ describe('merge-on-green', () => {
         getPRStub.restore();
       });
       it('does not add a PR if no branch protection', async () => {
-        //addPRStub.restore();
         loggerStub.restore();
 
         const scopes = [
@@ -883,7 +166,6 @@ describe('merge-on-green', () => {
       });
 
       it('adds a PR if branch protection when PR labeled', async () => {
-        //addPRStub.restore();
         const addPRStub = sandbox.stub(handler, 'addPR').callsFake(async () => {
           await handler.checkForBranchProtection(
             'testOwner',
@@ -911,7 +193,6 @@ describe('merge-on-green', () => {
 
       //This function is supposed to respond with an error
       it('does not add a PR if branch protection errors and comments on PR when PR labeled', async () => {
-        //addPRStub.restore();
         loggerStub.restore();
 
         const scopes = [getBranchProtection(400, []), commentOnPR()];
@@ -931,7 +212,7 @@ describe('merge-on-green', () => {
         scopes.forEach(s => s.done());
       });
     });
-    describe('stubbing all GCP functions', () => {
+    suite('stubbing all GCP functions', () => {
       let addPRStub: SinonStub;
       let removePRStub: SinonStub;
       let getPRStub: SinonStub;
@@ -1051,6 +332,7 @@ describe('merge-on-green', () => {
           assert(!addPRStub.called);
         });
       });
+
       describe('PRs when labeled', () => {
         handler.allowlist = ['testOwner'];
         it('adds a PR when label is added correctly', async () => {
@@ -1176,7 +458,7 @@ describe('merge-on-green', () => {
       });
     });
 
-    describe('PRs when closed, merged or unlabeled, testing getPR method (no stub)', () => {
+    suite('PRs when closed, merged or unlabeled, testing getPR method (no stub)', () => {
       let addPRStub: SinonStub;
       let removePRStub: SinonStub;
 
@@ -1447,4 +729,3 @@ describe('merge-on-green', () => {
       });
     });
   });
-});
